Enable two debug triggers (breakpoints) in artya7 FPGA example
Set the ibex_core parameters 'DbgTriggerEn' and 'DbgHwBreakNum' for
two hardware breakpoints when DBG is enabled.

Enabling this in the Ibex core is strongly desirable to present a good
debug experience.  Moreover, some RISC-V debug tools assume that
breakpoints exist on the target, so failure to have this in the example
only causes a frustrated user.